re-organize the codebase's ml dir, follow the same pattern as the other workers, with a generic worker class that needs to be implemented for each survey, and where each survey has it's own custom processing logic
