+++ docker-compose.yaml
@@ -1,13 +1,14 @@
 volumes:
   mongodb:
   valkey:
+  # kafka:
 
 networks:
   boom:
 
 services:
   mongo:
-    image: mongo:7.0
+    image: mongo:8.0
     hostname: mongo
     expose:
       - "27017"
@@ -32,7 +33,7 @@ services:
       - ALLOW_EMPTY_PASSWORD=yes
       - VALKEY_DISABLE_COMMANDS=FLUSHDB,FLUSHALL
     ports:
-      - '6379:6379'
+      - "6379:6379"
     volumes:
       - valkey:/data
     restart: always
@@ -46,7 +47,6 @@ services:
   broker:
     image: apache/kafka:latest
     hostname: broker
-    container_name: broker
     ports:
       - 9092:9092
     networks:
@@ -67,3 +67,6 @@ services:
       KAFKA_CONTROLLER_LISTENER_NAMES: CONTROLLER
       KAFKA_LOG_DIRS: /tmp/kraft-combined-logs
       KAFKA_NUM_PARTITIONS: 15
+    # volumes:
+    #   - ./data/kafka:/tmp/kraft-combined-logs
+    #   - kafka:/tmp/kraft-combined-logs
