+++ src/ml/mod.rs
@@ -1,2 +1,5 @@
 mod base;
-pub use base::run_ml_worker;
+mod models;
+mod ztf;
+pub use base::{run_ml_worker, MLWorker, MLWorkerError};
+pub use ztf::ZtfMLWorker;
